# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API documentation for use marketplace database.",
        "title": "Swagger Marketplace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6060",
    "basePath": "/api/v1",
    "paths": {
...
        "/me/": {
            "get": {
                "description": "get profile data from JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Get current user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
...
        "/users/:id": {
            "get": {
                "description": "get profile data from userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user profile by userID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "delete user data from userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Edit user profile by userID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete user data from userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Discard user profile by userID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "recovery_key": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.User"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}